Fail for untitled buffers
By implementing `run` we avoid running a subprocess to only return
a static string.  Since "echo" is a builtin on Windows this works across
the platforms as well.

Note that we also use "-" as the buffer name which is the canonical name
for "this buffer" or "stdin". SublimeLinter knows how to treat this
special name.